CORE-5120: Move `multipartDir`
