+++ libs/http-rpc/http-rpc-client/src/integrationTest/kotlin/net/corda/httprpc/client/HttpRpcIntegrationTestBase.kt
@@ -3,7 +3,6 @@ package net.corda.httprpc.client
 import net.corda.httprpc.server.HttpRpcServer
 import net.corda.httprpc.server.config.models.HttpRpcContext
 import net.corda.httprpc.test.utils.FakeSecurityManager
-import java.nio.file.Path
 
 abstract class HttpRpcIntegrationTestBase {
     internal companion object {
@@ -13,6 +12,5 @@ abstract class HttpRpcIntegrationTestBase {
         val userAlice = User("admin", password, setOf())
         val securityManager = FakeSecurityManager()
         val context = HttpRpcContext("1", "api", "HttpRpcContext test title ", "HttpRpcContext test description")
-        val multipartDir: Path = Path.of(System.getProperty("java.io.tmpdir"), "multipart")
     }
 }
