+++ libs/http-rpc/http-rpc-server-impl/src/integrationTest/kotlin/net/corda/httprpc/server/impl/utils/TestHelpers.kt
@@ -1,7 +1,3 @@
 package net.corda.httprpc.server.impl.utils
 
-import java.nio.file.Path
-
 fun String.compact() = this.trimMargin().replace("[\n\r]".toRegex(),"")
-
-internal val multipartDir = Path.of(System.getProperty("java.io.tmpdir"), "multipart")
