+++ libs/http-rpc/http-rpc-test-common/src/main/kotlin/net/corda/httprpc/test/utils/TestUtils.kt
@@ -1,5 +1,8 @@
 package net.corda.httprpc.test.utils
 
 import java.net.ServerSocket
+import java.nio.file.Path
 
 fun findFreePort() = ServerSocket(0).use { it.localPort }
+
+val multipartDir: Path = Path.of(System.getProperty("java.io.tmpdir"), "multipart")
